% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_surnames}
\alias{census_surnames}
\title{Frequently occurring surnames from the 2010 Census}
\format{A data frame with 162,254 rows and 11 variables:
\describe{
  \item{name}{name, in caps}
  \item{rank}{rank}
  \item{count}{raw count of name occurrence}
  \item{prop100k}{prevalence of name per 100,000}
  \item{cum_prop100k}{proportion of population, per 100k, having this name or a more common one}
  \item{pctwhite}{percent of people with this surname identified as White}
  \item{pctblack}{percent of people with this surname identified as Black}
  \item{pctapi}{percent of people with this surname identified as Asian, Native Hawaiian, and other Pacific Islander}
  \item{pctaian}{percent of people with this surname identified as American Indian and Alaska Native}
  \item{pct2prace}{percent of people with this surname identified as two or more races}
  \item{pcthispanic}{percent of people with this surname identified as hispanic/latino}
}}
\source{
\url{https://www.census.gov/topics/population/genealogy/data/2010_surnames.html}
}
\usage{
census_surnames
}
\description{
Tabulations of all surnames occurring 100 or more times in the 2010 Census.
}
\keyword{datasets}
