% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edit_counts}
\alias{edit_counts}
\title{Counts for single-edit spelling correction edits}
\format{A data framw ith 1,587 rows and 3 variables:
\describe{
  \item{from}{original character (single or two-character sequence)}
  \item{to}{replacement character (1- or 2-char sequence) after typo}
  \item{count}{frequency of this replacement in the typo corpus}
}}
\source{
\url{http://norvig.com/ngrams/}
}
\usage{
edit_counts
}
\description{
Common typos
}
\keyword{datasets}
