% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{name_dictionary}
\alias{name_dictionary}
\title{Hypocoristics for common English given names}
\format{A data frame with 755 rows and 2 variables:
\describe{
  \item{long}{long version of name}
  \item{short}{shortened version of name}
}}
\source{
\url{https://en.wiktionary.org/wiki/Appendix:English_given_names}
}
\usage{
name_dictionary
}
\description{
Dataset containing the shortened/alternate versions of names
that are common in English
}
\keyword{datasets}
