% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{surname_counts}
\alias{surname_counts}
\title{Surname counts by race, derived from Census data}
\format{A data frame with 162,254 rows and 11 variables:
\describe{
  \item{race}{race}
  \item{name}{name}
  \item{count}{calculated count based on overall name count and proportion by race}
}}
\source{
\url{https://www.census.gov/topics/population/genealogy/data/2010_surnames.html}
}
\usage{
surname_counts
}
\description{
Counts derived from \code{\link{census_surnames}} and names converted
to lower case.
}
\keyword{datasets}
